HW2/requirements.txt
torch>=2.11.0
torchvision>=0.26.0
scipy>=1.17.0
tqdm>=4.64.0
Pillow>=9.0.0
numpy>=2.0.0
tensorboard>=2.11.0
torchmetrics>=1.9.0
pycocotools>=2.0.11
